newrequirements.txt
ruamel.yaml
jdatetime
requests
pytz